[top_processor] Add plot_all script
This script is mean to be a single script to get the entirety of a
running appliance over a period of logs.  This is a wrapper script
around:

  - sync_dumps_and_logs (optional)
  - top_processor.rb
  - top_info_plot

By default it takes in an input of a `systemctl status evmserverd`
output, determines the relevant pids from those, and runs the above
scripts from those.

The process looks something like this:

* sync the dumps and logs from the server (if -s and -H is provided)
* process the pids from the file passed in from -f
* parse the top_output.log file(s) for every pid that would be in the
  `-f` file (anything with `MIQ Server`, `MIQ: `, or `puma ` in the
  proctitle)
* capture the generated files from the above, then for each pid, send
  the relavent file through `top_info_plot`
  - names are generated from the data in the -f file
* if -o is passed, open up all the files using `open`, otherwise output
  each graph that generated

A sample invocation of the command would look something like this:

    $ ../miq_tools/miq_top_output_processor/plot_all              \
        -siH 123.45.6.789 -p "CFME: " -u " with Pathnames Patch"  \
        -of evmserverd_snapshot_status.txt                        \
        top_output.log-[1,2,3,4]*